Montyformat datagen Bench: 3578727

+++ src/types.h
@@ -52,7 +52,7 @@ using std::cerr;
 using std::cout;
 using std::endl;
 
-enum Color : int {
+enum Color : i8 {
     WHITE = 1,
     BLACK = 0,
 };
@@ -88,7 +88,7 @@ enum PieceType : int {
 };
 
 // clang-format off
-enum Square : int {
+enum Square : i8 {
     a1, b1, c1, d1, e1, f1, g1, h1,
     a2, b2, c2, d2, e2, f2, g2, h2,
     a3, b3, c3, d3, e3, f3, g3, h3,
